moon/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
qrcode==7.4.2
reportlab==4.0.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Pillow==10.0.0
gunicorn==21.2.0
supabase-py==2.3.4
postgrest-py==0.15.1